Keep the same api as next/jest
