--- apps/web/jest.config.ts
@@ -1,7 +1,0 @@
-import type {Config} from "@jest/types"
-
-const config: Config.InitialOptions = {
-  preset: "@blitzjs/next/jest",
-}
-
-export default config
