+++ apps/web/tsconfig.json
@@ -1,8 +1,5 @@
 {
   "extends": "@blitzjs/config/tsconfig.nextjs.json",
-  "compilerOptions": {
-    // "isolatedModules": false
-  },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "jest.config.js"],
   "exclude": ["node_modules", "test"]
 }
