--- packages/blitz-next/jest/global-setup.js
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/no-default-export
-const loadEnvConfig = require("@blitzjs/env").loadEnvConfig
-module.exports = function globalSetup(globalConfig) {
-  loadEnvConfig()
-}
